Remove redundant moe test from test_models and use shared coordinate
